Clean up Public Controller
* Move app specific stuff to new public app service
* Create breathing room for NOIs and SRWs
